+++ frontend/src/components/ResultsView/ItemMenu/PersonCard/ItemRow/itemRow.component.tsx
@@ -1,11 +1,7 @@
 import { ChangeEvent, useEffect, useRef, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { didMount } from "../../../../../hooks";
-import {
-  AppStore,
-  Item,
-  SanitizedCurrency,
-} from "../../../../../models/custom-models";
+import { AppStore, Item, SanitizedCurrency } from "../../../../../models";
 import {
   editItemName,
   editItemPrice,