+++ frontend/src/models/person.model.ts
@@ -1,0 +1,10 @@
+export type Person = {
+  id: string;
+  name: string;
+  subtotal: string;
+  subtotalFloat: number;
+  tipAndTax: string;
+  tipAndTaxFloat: number;
+  totalDue: string;
+  totalDueFloat: number;
+};