+++ frontend/src/redux/calculation/initial.ts
@@ -1,4 +1,4 @@
-import { CalculationState } from "../../models/custom-models";
+import { CalculationState } from "../../models";
 import { v4 as uuidv4 } from "uuid";
 
 export const INITIAL_STATE: CalculationState = {