+++ CHANGELOG.md
@@ -51,3 +51,4 @@
 
  ## v0.2.1-beta 
  - [x] Display WebShareAPI only on viewports with less than 576px
+ - [x] Update components to export {default} and rename to naming pattern of .component, .service, .dto